examples: Use #[naked]
