+++ crates/examples/root-task/example-root-task-without-runtime/Cargo.nix
@@ -4,12 +4,11 @@
 # SPDX-License-Identifier: BSD-2-Clause
 #
 
-{ mk, localCrates, versions }:
+{ mk, localCrates }:
 
 mk {
   package.name = "example-root-task-without-runtime";
   dependencies = {
-    inherit (versions) cfg-if;
     sel4 = localCrates.sel4 // { default-features = false; };
   };
 }
