+++ crates/examples/root-task/example-root-task-without-runtime/Cargo.toml
@@ -17,5 +17,4 @@ edition = "2024"
 license = "BSD-2-Clause"
 
 [dependencies]
-cfg-if = "1.0.4"
 sel4 = { path = "../../../sel4", default-features = false }
