+++ crates/examples/root-task/example-root-task-without-runtime/src/main.rs
@@ -6,10 +6,10 @@
 
 #![no_std]
 #![no_main]
-#![feature(core_intrinsics)]
-#![allow(internal_features)]
 
-use core::arch::global_asm;
+mod runtime;
+
+enum Never {}
 
 fn main(bootinfo: &sel4::BootInfoPtr) -> sel4::Result<Never> {
     sel4::debug_println!("Hello, World!");
@@ -76,164 +76,4 @@ fn main(bootinfo: &sel4::BootInfoPtr) -> sel4::Result<Never> {
         .unwrap();
 
     unreachable!()
-}
-
-// minimal ad-hoc runtime
-
-enum Never {}
-
-mod stack {
-    use core::cell::UnsafeCell;
-
-    #[repr(C, align(16))]
-    pub struct Stack<const N: usize>(UnsafeCell<[u8; N]>);
-
-    unsafe impl<const N: usize> Sync for Stack<N> {}
-
-    impl<const N: usize> Stack<N> {
-        pub const fn new() -> Self {
-            Self(UnsafeCell::new([0; N]))
-        }
-
-        pub const fn bottom(&self) -> StackBottom {
-            StackBottom(self.0.get().cast::<u8>().wrapping_add(N))
-        }
-    }
-
-    #[repr(transparent)]
-    pub struct StackBottom(#[allow(dead_code)] *mut u8);
-
-    unsafe impl Sync for StackBottom {}
-
-    const STACK_SIZE: usize = 0x4000;
-
-    static STACK: Stack<STACK_SIZE> = Stack::new();
-
-    #[unsafe(no_mangle)]
-    static __stack_bottom: StackBottom = STACK.bottom();
-}
-
-cfg_if::cfg_if! {
-    if #[cfg(target_arch = "aarch64")] {
-        global_asm! {
-            r#"
-                .extern __rust_entry
-                .extern __stack_bottom
-
-                .section .text
-
-                .global _start
-                _start:
-                    ldr x9, =__stack_bottom
-                    ldr x9, [x9]
-                    mov sp, x9
-                    b __rust_entry
-
-                1:  b 1b
-            "#
-        }
-    } else if #[cfg(target_arch = "arm")] {
-            global_asm! {
-                r#"
-                    .extern __rust_entry
-                    .extern __stack_bottom
-
-                    .section .text
-
-                    .global _start
-                    _start:
-                        ldr r12, =__stack_bottom
-                        ldr r12, [r12]
-                        mov sp, r12
-                        b __rust_entry
-
-                    1:  b 1b
-                "#
-            }
-    } else if #[cfg(any(target_arch = "riscv64", target_arch = "riscv32"))] {
-        macro_rules! riscv_common {
-            () => {
-                r#"
-                    .extern __rust_entry
-                    .extern __stack_bottom
-
-                    .section .text
-
-                    .global _start
-                    _start:
-
-                        # See https://www.sifive.com/blog/all-aboard-part-3-linker-relaxation-in-riscv-toolchain
-                    .extern __global_pointer$
-                    .option push
-                    .option norelax
-                    1:  auipc gp, %pcrel_hi(__global_pointer$)
-                        addi gp, gp, %pcrel_lo(1b)
-                    .option pop
-
-                        la sp, __stack_bottom
-                        lx sp, (sp)
-                        jal __rust_entry
-
-                    1:  j 1b
-                "#
-            }
-        }
-
-        #[cfg(target_arch = "riscv64")]
-        global_asm! {
-            r#"
-                .macro lx dst, src
-                    ld \dst, \src
-                .endm
-            "#,
-            riscv_common!()
-        }
-
-        #[cfg(target_arch = "riscv32")]
-        global_asm! {
-            r#"
-                .macro lx dst, src
-                    lw \dst, \src
-                .endm
-            "#,
-            riscv_common!()
-        }
-    } else if #[cfg(target_arch = "x86_64")] {
-        global_asm! {
-            r#"
-                .extern __rust_entry
-                .extern __stack_bottom
-
-                .section .text
-
-                .global _start
-                _start:
-                    mov rsp, __stack_bottom
-                    mov rbp, rsp
-                    sub rsp, 0x8 // Stack must be 16-byte aligned before call
-                    push rbp
-                    call __rust_entry
-
-                1:  jmp 1b
-            "#
-        }
-    } else {
-        compile_error!("unsupported architecture");
-    }
-}
-
-#[unsafe(no_mangle)]
-unsafe extern "C" fn __rust_entry(bootinfo: *const sel4::BootInfo) -> ! {
-    let bootinfo = unsafe { sel4::BootInfoPtr::new(bootinfo) };
-    match main(&bootinfo) {
-        #[allow(unreachable_patterns)]
-        Ok(absurdity) => match absurdity {},
-        Err(err) => panic!("Error: {}", err),
-    }
-}
-
-#[panic_handler]
-fn panic(info: &core::panic::PanicInfo<'_>) -> ! {
-    sel4::debug_println!("{}", info);
-    core::intrinsics::abort()
 }
